style(user-page): add initial new ui
