+++ src/components/pages/HomePage.js
@@ -3,5 +3,5 @@ import React from 'react';
 import { UsernameSearchField } from '../../components/main/UsernameSearchField';
 
 export const HomePage = () => {
-  return <UsernameSearchField size="full" />;
+  return <UsernameSearchField />;
 };
